[Util] Refactor add months to epoch functions
