+++ src/bdap/fees.h
@@ -49,9 +49,6 @@ public:
 };
 
 bool GetBDAPFees(const opcodetype& opCodeAction, const opcodetype& opCodeObject, const BDAP::ObjectType objType, const uint16_t nMonths, CAmount& monthlyFee, CAmount& oneTimeFee, CAmount& depositFee);
-int64_t AddMonthsToCurrentEpoch(const short nMonths);
-int64_t AddMonthsToBlockTime(const uint32_t& nBlockTime, const short nMonths);
-uint16_t MonthsFromBlockToExpire(const uint32_t& nBlockTime, const uint64_t& nExpireTime);
 bool ExtractAmountsFromTx(const CTransactionRef& ptx, CAmount& dataAmount, CAmount& opAmount);
 
 #endif // DYNAMIC_BDAP_FEES_H
